server/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.6
django-cors-headers==4.3.0
django-filter==23.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
filelock==3.13.1
fsspec==2024.2.0
huggingface-hub==0.20.3
idna==3.6
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentencepiece==0.1.99
sqlparse==0.4.4
sympy==1.12
threadpoolctl==3.3.0
tokenizers==0.15.2
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
tqdm==4.66.2
transformers==4.37.2
typing_extensions==4.9.0
urllib3==2.2.0
